Move as subdirectory of manager and use plural naming convention
